first commit
Added a straight rip of the `HaskellMaterials` repo but modified README.
This will overall be cut down and modified to be more relevant to the
PureScript ecosystem which means we won't emphasize the same kinds of
application structure techniques but rather focus on common use cases
for PureScript application.

In `purerl` we may not actually need to use a lot of the same kinds of
types around application state that we are carrying around, etc.,
whereas we definitely do want to do so if we are using `Halogen`, for
example.

The dual contexts of PureScript here are important because they will
inform what we should be focusing on in terms of application structure
and design.